added onLoad function
